chore: Move GitSourcesService to componentconstructor package
