+++ internal/service/componentconstructor/gitsources.go
@@ -1,4 +1,4 @@
-package componentdescriptor
+package componentconstructor
 
 import (
 	"fmt"
